remove billing rates and recipients
